+++ src/app/word-add/word-add.component.html
@@ -1,2 +1,7 @@
   <div class="word-selector">
-    </div> 
+    <h3>или добавить свое слово</h3>
+    <input id='my-input2' type="text">{{name}}
+       <div style="visibility: hidden">Минимальная длина текста - 3 символа</div>
+       <ul id='my-url2' ></ul> <br />
+       <button class="btn btn-primary" >Добавить</button>
+  </div>
