+++ src/app/word-selector/word-selector.component.html
@@ -1,7 +1,7 @@
 <div class="word-selector">
   <h2>Выберите слова для изучения</h2>
   <ul>
-    <app-word-add > </app-word-add>
+    
     <li *ngFor="let item of selectedWords">
         <button type="button" class="close" aria-label="Close" (click)="removeWord(item.word)">
           <clr-icon aria-hidden="true" shape="close"></clr-icon>
@@ -11,11 +11,6 @@
     </ul>
   </div>    
   <button class="btn btn-primary" routerLink="/training" (click)="done()" [disabled]="selectedWords.length===0">Начать упражнение</button>
-  <div class="word-selector">
-    <h3>или добавить свое слово</h3>
-       <input id='my-input2' type="text">{{name}}
-          <div style="visibility: hidden">Минимальная длина текста - 3 символа</div>
-          <ul id='my-url2' ></ul> <br />
-          <button class="btn btn-primary" >Добавить</button>
-  </div> 
+  <app-word-add > </app-word-add>
+
