+++ src/app/word-selector/word-selector.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit, EventEmitter } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
 import allWords from './vocabulary';
 import { ShareService } from '../share.service';
